requirements.txt
langchain-google-genai
google-generativeai
claudette
pillow
replicate
PyYAML
python-fasthtml
transformers[torch]
scikit-learn
